move tickets to its own file
